+++ engine/access/rest/request/get_account_key_test.go
@@ -60,24 +60,24 @@ func Test_GetAccountKey_ValidParse(t *testing.T) {
 	err := getAccountKey.Parse(addr, keyIndex, height, chain)
 	assert.NoError(t, err)
 	assert.Equal(t, getAccountKey.Address.String(), addr)
-	assert.Equal(t, getAccountKey.Index, uint64(5))
+	assert.Equal(t, getAccountKey.Index, uint32(5))
 	assert.Equal(t, getAccountKey.Height, uint64(100))
 
 	err = getAccountKey.Parse(addr, keyIndex, "", chain)
 	assert.NoError(t, err)
 	assert.Equal(t, getAccountKey.Address.String(), addr)
-	assert.Equal(t, getAccountKey.Index, uint64(5))
+	assert.Equal(t, getAccountKey.Index, uint32(5))
 	assert.Equal(t, getAccountKey.Height, SealedHeight)
 
 	err = getAccountKey.Parse(addr, keyIndex, "sealed", chain)
 	assert.NoError(t, err)
 	assert.Equal(t, getAccountKey.Address.String(), addr)
-	assert.Equal(t, getAccountKey.Index, uint64(5))
+	assert.Equal(t, getAccountKey.Index, uint32(5))
 	assert.Equal(t, getAccountKey.Height, SealedHeight)
 
 	err = getAccountKey.Parse(addr, keyIndex, "final", chain)
 	assert.NoError(t, err)
 	assert.Equal(t, getAccountKey.Address.String(), addr)
-	assert.Equal(t, getAccountKey.Index, uint64(5))
+	assert.Equal(t, getAccountKey.Index, uint32(5))
 	assert.Equal(t, getAccountKey.Height, FinalHeight)
 }
